Fix pkg description
Per instructions from https://packaging.python.org/guides/making-a-pypi-friendly-readme/,
if we want to have a long_description from a README.md we need to set the
encoding on the read argument (even with python3), and we need to set a long_description_content_type.

With this fix, you could upload it to Pypi and the description would show up as expected.