Fix Client constructor for use on plaid-python 7.1.0 and plaid API version 2020-09-14
In the newer `plaid-python` library, the 3rd positional argument to the
`plaid.Client` constructor war removed [1], causing all of our other args to shift
and pass the token as the env, suppress warnings as the timeout, etc.

This updates the code to use kwargs instead, so that we don't depend on the
relative ordering in the class constructor. I've tested this successfully with
the versions listed in the subject line as part of packaging for NixPkgs [2]

With this commit, we can download new transactions with the latest libs and API,
which gives us most of the support for resolving #16 (just the docs/code on
creating new link tokens needs to be updated).

[1] https://github.com/plaid/plaid-python/blob/a0f7891dac304b6d8bee08ec4286fcaa8b5358f5/plaid/client.py#L39-L46
[2] https://github.com/NixOS/nixpkgs/pull/102368